requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Frozen-Flask==0.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dateutil==2.8.1
python-gitlab==1.13.0
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
